refactor: shorten chunked to chunk
